refactored backend folder
