+++ backend/__init__.py
@@ -1,12 +1,13 @@
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
 from flask_migrate import Migrate
 from config import Config
 
 app = Flask(__name__)
 app.config.from_object(Config)
 db = SQLAlchemy(app)
+login = LoginManager(app)
 migrate = Migrate(app, db)
 
-from backend.app import routes
-
+from backend import routes, models
