--- backend/app/routes.py
@@ -1,8 +1,0 @@
-from backend.app import app
-
-@app.route('/')
-@app.route('/index')
-def index():
-    return "Hello, World!"
-
-
